requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
langchain==0.1.4
langchain-openai==0.0.5
openai==1.12.0
pypdf==4.0.2
pillow==10.2.0
pytesseract==0.3.10
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0